% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvenienceFunctions.R
\name{detrend}
\alias{detrend}
\title{Detrend a vector}
\usage{
detrend(x, returntrend = FALSE)
}
\arguments{
\item{x}{Vector or time series to be detrended}

\item{returntrend}{When FALSE the detrended vector is
returned, when TRUE a 2 element list is returned with both the trend and the
detrended vector}
}
\value{
The detrended vector. When \code{returntrend = TRUE} a 2 element
list is returned with the trend and the detrended vector.
}
\description{
Substract a linear trend from a vector
}
\details{
The linear trend is calculated as \code{lm(x~1:length(x))}
}
\author{
Tom Cox <tom.cox@uantwerp.be>
}
\keyword{utilities}
