% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O22GPCompatibility.R
\name{O22GPpreprocess}
\alias{O22GPpreprocess}
\alias{O22GP}
\alias{WindowO22GP}
\alias{WindowO22GP.gts}
\alias{WindowO22GP}
\alias{WindowO22GP.gts}
\title{Deprecated: Calculate GPP from O2 time series}
\usage{
O22GPpreprocess(...)

O22GP(...)

WindowO22GP(...)

WindowO22GP.gts(...)
}
\arguments{
\item{...}{All arguments are passed to equivalent GPPFourier functions}
}
\value{
See \code{\link{GPPFourier}}, \code{\link{GPPFourierPreprocess}},
\code{\link{WindowGPPFourier}}, \code{\link{WindowGPPFourier.gts}}
}
\description{
Don't use this function, use GPPFourier() instead. This function only exists
for backward compatibility of GPPFourier package with older scripts.
}
\author{
Tom Cox <tom.cox@uantwerp.be>
}
\keyword{utilities}
