% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourierFunctions.R
\name{paddedfft}
\alias{paddedfft}
\title{Fourier transform of padded time series}
\usage{
paddedfft(x, n)
}
\arguments{
\item{x}{Time series for which fourier transform (fft) is calculated after
padding with zeroes}

\item{n}{Number of zeroes to be padded}
}
\value{
Fourier transform of x padded with n zeroes
}
\description{
Fourier transform of padded time series
}
\author{
Tom Cox <tom.cox@uantwerp.be>
}
\keyword{#}
\keyword{@export}
\keyword{internal}
\keyword{paddedfft}
\keyword{utilities}
