% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourierFunctions.R
\name{padTestPlot}
\alias{padTestPlot}
\title{Test plot for padded fft calculation}
\usage{
padTestPlot(y, padlength, dt = 1, ...)
}
\arguments{
\item{y}{Vector or time series for which fft and padded fft is calculated}

\item{padlength}{Number of zeroes to be appended to \code{y}}

\item{dt}{Time interval between succesive values of \code{y}}

\item{...}{Other arguments passed to plot()}
}
\value{
A two panel plot is generated with the unpadded fft on top and the
padded fft at the bottom
}
\description{
Test plot for padded fft calculation
}
\author{
Tom Cox <tom.cox@uantwerp.be>
}
\keyword{#@export}
\keyword{internal}
\keyword{padTestPlot}
\keyword{utilities}
